Rpath: A package implementing mass balance algorithms designed to work with fisheries data sources.

The Rpath package provides two categories of important functions: rpath and rsim.

The Rpath package provides two categories of important functions: rpath and rsim.

Rpath functions

The Rpath functions generate the balanced snap shot of energy flow through the system. This is done by solving a series of linear equations for unknown biomass or ecotrophic efficiency.

rpath functions

The rpath functions generate the balanced snap shot of energy flow through the system. This is done by solving a series of linear equations for unknown biomass or ecotrophic efficiency.

See also

Author

Maintainer: Kerim Aydin Kerim.Aydin@NOAA.gov

Authors:

Sean Lucey Sean.Lucey@NOAA.gov

Sarah Gaichas Sarah.Gaichas@NOAA.gov

Sarah Weisberg sarah.j.weisberg@stonybrook.edu

Andy Whitehouse Andy.Whitehouse@NOAA.gov

Other contributors:

Ron Klasky ronald.klasky@noaa.gov [contributor]

Andy Beet andrew.beet@noaa.gov [contributor]

Max Grezlik maxwell.grezlik@noaa.gov [contributor]