Adjust Fishing Mortality

Modifies the fishing mortality value for a species by a particular gear. Parameters that can be adjusted using this function are: ForcedEffort, ForcedFRate, or ForcedCatch.

Usage

adjust.fishing(
  Rsim.scenario,
  parameter,
  group = NA,
  sim.year = 1,
  sim.month = 0,
  value
)

Arguments

Rsim.scenario

Scenario object that contains all of the rsim rates and forcing functions generated by rsim.scenario().

parameter

Parameters to be modified (ForcedEffort,ForcedCatch,ForcedFRate)

group

Name of the group whose parameter is being changed. Valid values are found in the `Group` field of the object created from running rpath()

sim.year

Year of the simulation that should be modified. Can be a range of years.

sim.month

Month of the year that should be modified. If set to 0, all months of the year are modified.

value

New value for the parameter.

Value

Returns an Rsim.scenario() object with the new fishing parameter values.

See also

Examples

# Read in Rpath parameter file and generate balanced model
Rpath <- rpath(AB.params)
# Create a 50 yr Rsim scenario
Rsim.scenario <- rsim.scenario(Rpath, AB.params, years = 1:50)
# Change value of forcedFRate for Squids in years 3 through 5 to the value of 2 (for all months)
Rsim.scenario.adjusted.fishing <- adjust.fishing(Rsim.scenario,parameter="ForcedFRate",group="cod",sim.year=3:5,value = 2)
head(Rsim.scenario.adjusted.fishing$fishing$ForcedFRate)    
#>   Outside whales seals cod whiting mackerel anchovy shrimp benthos zooplankton
#> 1       0      0     0   0       0        0       0      0       0           0
#> 2       0      0     0   0       0        0       0      0       0           0
#> 3       0      0     0   2       0        0       0      0       0           0
#> 4       0      0     0   2       0        0       0      0       0           0
#> 5       0      0     0   2       0        0       0      0       0           0
#> 6       0      0     0   0       0        0       0      0       0           0
#>   phytoplankton detritus
#> 1             0        0
#> 2             0        0
#> 3             0        0
#> 4             0        0
#> 5             0        0
#> 6             0        0