Adjust Forcing Parameters

Modifies the various forcing parameters of the rsim scenario object.

adjust.forcing(
  Rsim.scenario,
  parameter,
  group,
  sim.year = 1,
  sim.month = 0,
  bymonth = F,
  value
)

Arguments

Rsim.scenario

Object generated by rsim.scenario.

parameter

The Rsim.scenario forcing parameter to be modified.

group

The model group that the parameter change will affect.

sim.year

Year of the simulation that should be modified. Can be a range of years.

sim.month

Month of the year that should be modified. If set to 0, all months of the year are modified.

bymonth

Boolean value that denotes whether to use sim.year/sim.month combo or just sim.month as a sequential vector starting at 1.

value

The new value for the parameter.

Value

Returns an Rsim.scenario object with the new parameter.

See also