Fishing Mortality Table

Creates a table of fishing mortalities by species group and gear for an rsim.scenario() object.

Usage

frate.table(Rsim.scenario)

Arguments

Rsim.scenario

Scenario object that contains all of the rsim rates and forcing functions generated by rsim.scenario().

Value

Returns a data table of F values for each species/gear combination.

Group

Group/node

Fishery/gearType

Gear type for the fishery

...

Total

Sum over all gear types

See also

Examples

# Read in Rpath parameter file and generate balanced model
Rpath <- rpath(AB.params)
# Create a 50 yr Rsim scenario
Rsim.scenario <- rsim.scenario(Rpath, AB.params, years = 1:50)
rates <- frate.table(Rsim.scenario)
# display the head of the data frame
head(rates)
#>       Group    sealers  trawlers   seiners bait boats shrimpers      Total
#>      <char>      <num>     <num>     <num>      <num>     <num>      <num>
#> 1:    seals 0.07389163 0.0000000 0.0000000 0.00000000    0.0000 0.07389163
#> 2:      cod 0.00000000 0.1500000 0.0000000 0.00000000    0.0000 0.15000000
#> 3:  whiting 0.00000000 0.1111111 0.0000000 0.00000000    0.0000 0.11111111
#> 4: mackerel 0.00000000 0.0000000 0.3333333 0.00000000    0.0000 0.33333333
#> 5:  anchovy 0.00000000 0.0000000 0.1714286 0.02857143    0.0000 0.20000000
#> 6:   shrimp 0.00000000 0.0000000 0.0000000 0.00000000    0.0625 0.06250000