Retrieve fishing forcing parameters from an Rsim scenario

Helper function that will retrieve the fishing forcing parameters that were used in an Rsim scenario

Usage

get.rsim.fishing(Rsim.scenario)

Arguments

Rsim.scenario

Scenario object that contains all of the rsim rates and forcing functions generated by rsim.scenario().

Value

Returns a `fishing` object.

See also

Examples

# Read in Rpath parameter file and generate balanced model
Rpath <- rpath(AB.params)
# Create a 50 yr Rsim scenario
Rsim.scenario <- rsim.scenario(Rpath, AB.params, years = 1:50)
params <- get.rsim.fishing(Rsim.scenario)
names(params)
#> [1] "ForcedEffort" "ForcedFRate"  "ForcedCatch"