Ecopath module of Rpath

Performs initial mass balance using a model parameter file and diet matrix file.

Usage

rpath(Rpath.params, eco.name = NA, eco.area = 1)

Arguments

Rpath.params

R object containing the Rpath parameters. This is generated either by the create.rpath.params or read.rpath.params functions.

eco.name

Optional name of the ecosystem which becomes an attribute of rpath object.

eco.area

Optional area of the ecosystem which becomes an attribute of the rpath object.

Value

Returns an Rpath object that can be supplied to the rsim.scenario function.

See also