Rpath functional group names

Get a character vector of functional group names from an Rpath object (balanced model) or an Rpath.params object (unbalanced model parameters) based on the 'type' input parameter as follows: (0: consumers, 1: producers, 2: detrital, 3: gears, 0<type<1: mixotrophs). Living groups are consumers + producers. Note that mixotrophs are not returned as either consumers or producers, only separately.

Usage

rpath.groups(Rpath)

Arguments

Rpath

Balanced Rpath model generated by rpath()

Value

Returns a string vector containing the names of Rpath functional groups by category (group type).

See also

Examples

# Build the balanced Rpath model and parameter file by calling `rpath`
Rpath <- rpath(Ecosense.EBS)
# Run the groups command
groups = rpath.groups(Rpath)
# Print out the first few group names
head(groups)
#> [1] "Toothed whales"      "Gray whales"         "Other baleen"       
#> [4] "Bowhead whales"      "Other pinnipeds"     "Walrus/bearded seal"