Calculate biomass and consumption for multistanza groups

Uses the leading stanza to calculate the biomass and consumption of other stanzas necessary to support the leading stanza.

Usage

rpath.stanzas(Rpath.params)

Arguments

Rpath.params

R object containing the Rpath parameters. This is generated either by the create.rpath.params or read.rpath.params functions.

Value

Calculates and adds biomass and consumption for trailing stanza groups. Also adds weight at age and number at age for multi-staza groups.

See also