Initialize forcing matrix for Rsim scenario

Creates a matrix for forcing functions not related to fishing (ForcedPrey ForcedMort, ForcedRecs, ForcedSearch, ForcedActResp, ForcedMigrate, and ForcedBio). Default values are set to 1 for ForcedPrey ForcedMort, ForcedRecs, ForcedSearch and ForcedActResp, to 0 for ForcedMigrate and to -1 for ForcedBio for all timesteps and groups.

Usage

rsim.forcing(params, years)

Arguments

params

Rsim parameter object generated by rsim.params.

years

A vector of each year of the simulation.

Value

S3 object class Rsim.forcing, a list of 7 objects:

ForcedPrey, numeric matrix of monthly (rows) prey forcing by biomass group (columns)

ForcedMort, numeric matrix of monthly (rows) mortality multiplier by biomass group (columns)

ForcedRecs, numeric matrix of monthly (rows) recruitment multiplier by biomass group (columns)

ForcedSearch, numeric matrix of monthly (rows) search rate multiplier by biomass group (columns)

ForcedActresp, numeric matrix of monthly (rows) active respiration multiplier by biomass group (columns)

ForcedMigrate, numeric matrix of monthly (rows) migration forcing by biomass group (columns)

ForcedBio, numeric matrix of monthly (rows) biomass multiplier by biomass group (columns)

Examples

# Read in Rpath parameter file and generate model object
Rpath <- rpath(AB.params)
# Create dynamic parameters from Rpath model
Rsim.params <- rsim.params(Rpath)
# Create forcing matrix with default values
Rsim.forcing <- rsim.forcing(Rsim.params, years = 1:50)