Plot routine for Rsim runs

Plots the relative biomass of each group from an Rsim.output object.

Usage

rsim.plot(Rsim.output, spname = "all", indplot = F)

Arguments

Rsim.output

R object containing the output from rsim.run().

spname

Vector of species names to be displayed. The default "all" will display all group names

indplot

Logical value of whether to plot a single group or multiple groups

Value

Creates a figure of relative biomass

Details

For a detailed example of how to use this plot command please see vignette("RunRsim")

See also