Specify a dynamic simulation scenario

Uses a static Rpath model and creates a scenario consisting of 5 list objects: params generated by rsim.params(), start_state generated by rsim.state(), forcing generated by rsim.forcing(), fishing generated by rsim.fishing(), and 'stanzas generated by rsim.stanzas().

Usage

rsim.scenario(Rpath, Rpath.params, years = 1:100)

Arguments

Rpath

R object containing a static Rpath model.

Rpath.params

R object containing the parameters needed to create a Rpath model. This is generated either by the create.rpath.params() or read.rpath.params() functions.

years

A vector of each year of the simulation.

Value

Returns an Rsim.scenario object that can be supplied to the rsim.run() function. Rsim.scenario is a list of 5 objects:

params, list of 39 dynamic parameters and vectors derived from an Rpath model, see output of rsim.params()

start_state, list of 9 initial value vectors for biomass and size structured stanza groups, see output of rsim.state() and rsim.stanzas()

forcing, list of 7 matrices for time series forcing of biological attributes, see output of rsim.forcing()

fishing, list of 3 matrices for time series forcing of fishing attributes, see output of rsim.fishing()

stanzas, list of 21 parameters and vectors for age structured groups, see output of rsim.stanzas()

See also

Examples

# Read in Rpath parameter file and generate model object
Rpath <- rpath(AB.params)
# Create a 50 yr Rsim scenario
Rsim.scenario <- rsim.scenario(Rpath, AB.params, years = 1:50)