Ecosense function for rpath (rsim.sense)

Kerim Aydin 31 December 2019

Usage

rsim.sense(Rsim.scenario, Rpath.params, Vvary = c(0, 0), Dvary = c(0, 0))

Arguments

Rsim.scenario

Rsim scenario object used as center of distributions (base model) during random parameter generation.

Rpath.params

Rpath parameter object (unbalanced model) used for data pedigree input.

Vvary

length-2 vector with (lower,upper) bounds of vulnerability generation in log-space - 1, Vvary = log(X-1) so Walters et al. 1997 range of (1..inf) centered on 2 becomes (-inf,+inf) centered on 0.

Dvary

length-2 vector with (lower,upper) bounds of handling time generation in log-space - 1, scaled as Vvary, above.

Value

Returns an Rsim.params object that can be substituted for the params in an Rsim.scenario object.