Function to add a time step to an Rsim run

Runs rsim.run and appends the output to a previous Rsim.output. Currently only works with the Adams-Bashforth numerical integration method (AB).

Usage

rsim.step(Rsim.scenario, Rsim.output, method = "AB", year.end)

Arguments

Rsim.scenario

Scenario object that contains all of the rsim rates and forcing functions generated by rsim.scenario.

Rsim.output

R object containing the output from rsim.run.

method

Numerical integration method. Either 'AB' for Adams-Bashforth or 'RK4' for 4th order Runge-Kutta.

year.end

The final year for the step interval

Value

Returns an Rsim.output object.

See also