Plot routine for Ecopath multistanzas

Plots the biomass composition of multistanza groups from an Rpath.stanzas object.

stanzaplot(
  Rpath.params,
  StanzaGroup,
  line.cols = c("black", "green", "blue", "red")
)

Arguments

Rpath.params

R object containing the Rpath parameters. This is generated either by the create.rpath.params or read.rpath.params functions.

StanzaGroup

The Stanza group's name to be plotted.

line.cols

A vector of four colors used to represent the population biomass, relative number, indvidual weights, and stanza separation lines.

Value

Creates a figure showing the break down of biomass and number per stanza.

See also