Ecosense function for Rpath

rsim.sense() draws a random Rsim parameter set from distributions centered on the supplied Rpath parameter estimates. The width of the distribution for each parameter is defined by the data pedigree of the Rpath.params object generated by either create.rpath.params() or read.rpath.params(). The returned Rsim parameter set is functionally equivalent to a params list object created by rsim.scenario().

Usage

rsim.sense(Rsim.scenario, Rpath.params, Vvary = c(0, 0), Dvary = c(0, 0))

Arguments

Rsim.scenario

The rsim.scenario() object containing the initial rsim parameters used as the center of distributions (base model) during random parameter generation.

Rpath.params

Rpath parameter object (unbalanced model) containing the data pedigree generated by either create.rpath.params() or read.rpath.params().

Vvary

length-2 vector with (lower,upper) bounds of vulnerability parameters generated in log-space - 1, Vvary = log(X-1). Thus, the Walters et al. 1997 vulnerability range of (1,+inf) centered on 2 now becomes (-inf,+inf) centered on 0.

Dvary

length-2 vector with (lower,upper) bounds of handling time generation in log-space - 1. Scaled as Vvary, above (-inf,+inf), centered on 0.

Value

Returns a params list object that can be substituted for the params list in an rsim.scenario object.

References

Walters C, Christensen V, Pauly D (1997) Structuring dynamic models of exploited ecosystems from trophic mass-balance assessments. Rev Fish Biol Fish 7:139-172. doi:10.1023/a:1018479526149

Whitehouse GA, Aydin KY (2020) Assessing the sensitivity of three Alaska marine food webs to perturbations: an example of Ecosim simulations using Rpath. Ecol Model 429:16. doi:10.1016/j.ecolmodel.2020.109074

See also

Examples

# Read in Rpath parameter file and balance model
EBS_bal <- rpath(Ecosense.EBS)
# Create a 100 yr Rsim scenario
EBS_scene <- rsim.scenario(EBS_bal, Ecosense.EBS, years = 1:100)
# Generate a new params list
new_EBS_params <- rsim.sense(EBS_scene, Ecosense.EBS, Vvary = c(-4.5, 4.5), Dvary = c(-4.5, 4.5))
# assign the generated params to the Rsim.scenario object
EBS_scene$params <- new_EBS_params