Plot routine for Ecopath food web

Plots the food web associated with an Rpath object.

webplot(
  Rpath.obj,
  eco.name = attr(Rpath.obj, "eco.name"),
  line.col = "grey",
  highlight = NULL,
  highlight.col = c("black", "red", "orange"),
  labels = FALSE,
  label.pos = NULL,
  label.num = FALSE,
  label.cex = 1,
  fleets = FALSE,
  type.col = "black",
  box.order = NULL
)

Arguments

Rpath.obj
Rpath model created by the ecopath() function.

eco.name
Optional name of the ecosystem. Default is the eco.name attribute from the rpath object.

line.col
The color of the lines between nodes of the food web.

highlight
Set to the group number or name to highlight the connections of that group.

highlight.col
Color of the connections to the highlighted group.

labels
Logical whether or not to display group names. If True and label.pos is Null, no points will be ploted, just label names.

label.pos
A position specifier for the labels. Values of 1, 2, 3, 4, respectively indicate positions below, to the left of, above, and to the right of the points. A null value will cause the labels to be ploted without the points (Assuming that labels = TRUE).

label.num
Logical value indication whether group numbers should be used for labels instead of names.

label.cex
The relative size of the labels within the plot.

fleets
Logical value indicating whether or not to include fishing fleets in the food web.

type.col
The color of the points cooresponding to the types of the group. Can either be of length 1 or 4. Color order will be living, primary producers, detrital, and fleet groups.

box.order
Vector of box numbers to change the default plot order. Must include all box numbers

Value

Creates a figure of the food web.

See also